make the tool look good (#32)
* remove duplicate route in main.js

* add vue-json-to-csv dependency

* add csv export button to tables

* add vuetify package

* conver transaction page to v-data-table

This commit switches the transactions page to using a Vuetify table component.

* convert nav bar into a nav drawer element

* add app bar component

* add routing for new pages

* remove links to work-in-progress pages

* convert rules table to v-card component

* make the table buttons look good

* move AppBar.vue to components directory

* add logo to login page

* add links to social media icons

* refine app bar

* wrap content in v-main tag

This commit wraps page content in the <v-main> tag. This is required by
Vuetify in order to properly render content.

* add dark mode

This commit adds the necessary support for dark mode. It adds a switch
components, inserts that component in the nav bar, and defines light/dark
color schemes in Vuetify.

* remove help and notifications pages

Team decided to remove these pages during last meeting. This commit does so.

* change login redirect to transactions page

* refine app bar icons

* create account page

* make transaction loading message reactive

This commit modifies the data table on the transactions page so that the
loading message is hidden after load.

* format data on transactions page

This commit formats time and currency values on the Transactions data table.

* add color chips to transaction table

* add icon to transaction table title

* fix button css

* fix edit action on rules table

* fix new rule creation

* fix new transaction page

* put new rule page routing back

Fixing a bug caught by Jake. Somehow the routing path for the new rule
page got removed.

* fix rule creation

This commit corrects an error (empty input) preventing new rule creation.
I also corrected a couple of CSS issues.

* change name on account page

* fix column header typo